+++ source/B2B.Transactions.IntegrationTests/TestBase.cs
@@ -45,7 +45,7 @@ namespace B2B.Transactions.IntegrationTests
                     correlation.SetId(Guid.NewGuid().ToString());
                     return correlation;
                 })
-                .AddMessagePublishing(new OutgoingMessageStoreSpy(), new DataAvailableNotificationSenderSpy());
+                .AddMessagePublishing(new OutgoingMessageStoreSpy(), new DataAvailableNotificationPublisherSpy());
             _serviceProvider = services.BuildServiceProvider();
         }
